expand vault capabilities field width for forward compatibility
Increase the vault plaintext capabilities type from u8 to u32 so capability flags have room to grow without overflow.

Made-with: Cursor

+++ crates/zode-app/src/vault.rs
@@ -17,7 +17,7 @@ pub(crate) struct VaultPlaintext {
     pub identity_id: [u8; 16],
     pub machine_id: [u8; 16],
     pub epoch: u64,
-    pub capabilities: u8,
+    pub capabilities: u32,
     pub libp2p_keypair: Vec<u8>,
 }
